Merge: 빌드 에러 해결

See merge request s09-fintech-finance-sub2/S09P22B207!183

+++ Back-End/src/main/java/com/conseller/conseller/notification/NotificationServiceImpl.java
@@ -342,7 +342,7 @@ public class NotificationServiceImpl implements NotificationService{
     }
 
     @Override
-    public void sendGifticonNotification(Long userIdx, Integer remainDay, String gifticionName, Integer gifticonCount, Integer type) {
+    public void sendGifticonNotification(Long userIdx, Integer remainDay, String gifticionName, Long gifticonCount, Integer type) {
         User user = userRepository.findById(userIdx)
                 .orElseThrow(() -> new CustomException(CustomExceptionStatus.GIFTICON_INVALID));
